config_tools: generate vDSDT according to allocated resources
This patch generates vDSDT for pre-launched VMs based on the scenario
configuration and static allocation results. The vDSDT contains the
following objects.

  * A device object representing the root PCI bus which has the same
    location and name as the object in physical DSDT. This object contains
    device identification objects such as _HID, _CID, _BBN and _UID, the
    _CRS object encoding the bus numbers and PCI hole regions, and the _PRT
    object encoding the routing from vPCI device pins to virtual interrupt
    lines.

  * For each pass-through device assigned to the VM, a device object is
    created based on the AML template of that device in board XML (if
    exists). The _ADR object is rewritten with the vBDF..

v1 -> v2:
  * In order to replace the current static vDSDT, the AML templates in
    board XMLs now include objects that are referenced by other devices as
    well.
  * When TPM is assigned, the corresponding device object in DSDT (if
    exists) will be copied.
  * Add the _S5 control method to vDSDT, as is done in the static DSDT
    sources.
  * The old vDSDT generation routine and the static DSDT sources it uses
    are removed.

v2 -> v3:
  * Only follow 'uses', 'is used by' or 'consumes resources by'
    dependencies.
  * Organize the device objects in vDSDT according to the scopes they are
    in.

Tracked-On: #6287

+++ misc/config_tools/acpi_gen/bin_gen.py
@@ -42,15 +42,19 @@ def asl_to_aml(dest_vm_acpi_path, dest_vm_acpi_bin_path):
                 shutil.copyfile(os.path.join(dest_vm_acpi_path, acpi_table[0]),
                                 os.path.join(dest_vm_acpi_bin_path, acpi_table[1]))
         else:
-            rc = exec_command('iasl {}'.format(acpi_table[0]))
-            if rc == 0 and os.path.isfile(os.path.join(dest_vm_acpi_path, acpi_table[1])):
-                shutil.move(os.path.join(dest_vm_acpi_path, acpi_table[1]),
+            if acpi_table[0].endswith(".asl"):
+                rc = exec_command('iasl {}'.format(acpi_table[0]))
+                if rc == 0 and os.path.isfile(os.path.join(dest_vm_acpi_path, acpi_table[1])):
+                    shutil.move(os.path.join(dest_vm_acpi_path, acpi_table[1]),
+                                os.path.join(dest_vm_acpi_bin_path, acpi_table[1]))
+                else:
+                    if os.path.isfile(os.path.join(dest_vm_acpi_path, acpi_table[1])):
+                        os.remove(os.path.join(dest_vm_acpi_path, acpi_table[1]))
+                    rmsg = 'failed to compile {}'.format(acpi_table[0])
+                    break
+            elif acpi_table[0].endswith(".aml"):
+                shutil.copy(os.path.join(dest_vm_acpi_path, acpi_table[0]),
                             os.path.join(dest_vm_acpi_bin_path, acpi_table[1]))
-            else:
-                if os.path.isfile(os.path.join(dest_vm_acpi_path, acpi_table[1])):
-                    os.remove(os.path.join(dest_vm_acpi_path, acpi_table[1]))
-                rmsg = 'failed to compile {}'.format(acpi_table[0])
-                break
 
     os.chdir(curr_path)
     if not rmsg:
